+++ src/components/ThroughputChart/index.ts
@@ -1,4 +1,6 @@
 export { ThroughputChart } from "./ThroughputChart";
-export type { ThroughputChartProps, ThroughputPoint } from "./ThroughputChart";
-export { CompletionChart } from "./CompletionChart";
-export type { CompletionChartProps, CompletionPoint } from "./CompletionChart";
+export type {
+  ThroughputChartProps,
+  ThroughputPoint,
+  CompletionPoint,
+} from "./ThroughputChart";